DRTII-1686 Add route & test
