+++ src/test/resources/.~lock.bx-example.xlsx#
@@ -1,0 +1,1 @@
+,rich,fcl,12.12.2024 12:27,file:///home/rich/.config/libreoffice/4;
